Fix issue command
